Fixing bug & improvement
- Avoid mem leak in glfw
- Fix Build warnings to do with BufferElement Offset Type
- Convert smart pointers to Chromaria wrappers: (Create)Scope and (Create)Ref
- Removing all DLL traces (removing all CHROMARIA_API macros)
- Fixing include statements
- Use default (de)constructor instead of an empty one
- Removed the dedicated (duplicate) `BIND_EVENT_FN` in application
- Call Renderer2D::Shutdown() to free memory from allocated Renderer2DStorage
- Removed traces of OpenGL in the main Renderer and in Sandbox/ExampleLayer

+++ Chromaria/src/Chromaria/Renderer/Buffer.h
@@ -33,7 +33,7 @@ namespace Chromaria{
 		std::string Name;
 		ShaderDataType Type;
 		uint32_t Size;
-		uint32_t Offset;
+		size_t Offset;
 		bool Normalized;
 
 		BufferElement(ShaderDataType type, const std::string& name, bool normalized = false)
@@ -84,7 +84,7 @@ namespace Chromaria{
 	private:
 		void CalculateOffsetAndStride()
 		{
-			uint32_t offset = 0;
+			size_t offset = 0;
 			m_Stride = 0;
 			for (auto& element : m_Elements)
 			{
